Fix issue with bucket names with dots in them
requests raises an error for buckets like www.example.com, because
when it tries to fetch a https://www.example.com.s3.amazonaws.com/
for which it gets invalid certificates, because Amazon issues certs
only for *.s3.amazonaws.com, but not for *.*.*.s3.amazonaws.com.

This was already handled before, but lost in migration to lambda
function, so added it back.